feat: fetch item price

+++ Assets/Scripts/Scriptable Object/ShopItemData.cs
@@ -6,6 +6,6 @@ public class ShopItemData : ScriptableObject
     public Sprite icon;
     public string id;
     public string itemName;         
-    public float price;
+    public decimal price;
     [TextArea] public string info; 
 }
